Ante 228 refactorunitteststypescript

+++ contracts/mock/AnteOddBlockTest.sol
@@ -14,11 +14,11 @@ pragma solidity ^0.7.0;
 import "../AnteTest.sol";
 
 /// Ante Test to check if the block height is odd (lol)
-contract AnteOddBlockTest is AnteTest("Toggleable Ante Odd Block Test") { 
-    uint public lastUpdate;
+contract AnteOddBlockTest is AnteTest("Toggleable Ante Odd Block Test") {
+    uint256 public lastUpdate;
     bool public willTest = false;
 
-    constructor () {
+    constructor() {
         protocolName = "OddBlock";
     }
 
@@ -26,11 +26,10 @@ contract AnteOddBlockTest is AnteTest("Toggleable Ante Odd Block Test") {
         willTest = _willTest;
     }
 
-
     function checkTestPasses() public override returns (bool) {
         if (willTest) {
             lastUpdate = block.number;
-            return (block.number %2 == 1);
+            return (block.number % 2 == 1);
         } else {
             return true;
         }
